Turning bo-tie into a generic library for bluetooth
This requires that the HCI allows for an implementation of an interface to interact with the controller. For platforms (such as linux) that
already implement the blueooth controller, there just needs to be a simple implementation.

All of the panics in the HCI events module have been replace for returning an error (with the error type being String)