Added basic functionality for both client & server of the Bluetooth Attribute Protocol.
Right now the server can only handle the MTU Request, Read Request, and Write Request PDU.
These were lightly tested and may still contain problems.

The Client has methods to send more then those requests, but they will cause a panic on the server side.

The unit tests in the src/hci/mod.rs file were removed to the lib/linux.xrs file as they depend on linux and cannot be run on all systems. They are not compilable as a result of no longer being in the same crate. The file will be renamed to the proper rust extension when they work again.